+++ plutarch-orphanage/CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## 1.0.3 -- 04-07-2024
+
+### Added
+
+* `MintValue` as a helper for non-Ada `Value`s that are nonzero
+
+### Fixed
+
+* `FeeValue` now produces positive amounts only (as it should have from the
+  get-go).
+
 ## 1.0.2 -- 27-06-2024
 
 ### Added